pyproject.toml
[project]
name = "usa-agriculture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fpdf2>=2.8.5",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pydeck>=0.9.1",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
    "streamlit-lottie>=0.0.5",
]
